+++ tests/test_chunk_read_write.py
@@ -9,6 +9,17 @@ import pytest
 
 import laspy
 
+def check_chunked_reading_gives_expected_points(
+    groundtruth_las: laspy.LasData, reader: laspy.LasReader, iter_size: int
+):
+    """Checks that the points read by the reader are the same as groundtruth points."""
+    assert groundtruth_las.point_format == reader.header.point_format
+    for i, points in enumerate(reader.chunk_iterator(iter_size)):
+        expected_points = groundtruth_las.points[i * iter_size : (i + 1) * iter_size]
+        for dim_name in points.array.dtype.names:
+            assert np.allclose(
+                expected_points[dim_name], points[dim_name]
+            ), f"{dim_name} not equal"
 
 def test_chunked_las_reading_gives_expected_points(las_file_path):
     """
@@ -17,8 +28,7 @@ def test_chunked_las_reading_gives_expected_points(las_file_path):
     with laspy.open(las_file_path) as las_reader:
         with laspy.open(las_file_path) as reader:
             las = las_reader.read()
-            check_chunked_reading_is_gives_expected_points(las, reader, iter_size=50)
-
+            check_chunked_reading_gives_expected_points(las, reader, iter_size=50)
 
 def test_chunked_laz_reading_gives_expected_points(laz_file_path, laz_backend):
     """
@@ -27,7 +37,7 @@ def test_chunked_laz_reading_gives_expected_points(laz_file_path, laz_backend):
     with laspy.open(laz_file_path) as las_reader:
         with laspy.open(laz_file_path, laz_backend=laz_backend) as laz_reader:
             expected_las = las_reader.read()
-            check_chunked_reading_is_gives_expected_points(
+            check_chunked_reading_gives_expected_points(
                 expected_las, laz_reader, iter_size=50
             )
 
@@ -59,23 +69,9 @@ def test_chunked_writing_gives_expected_points(file_path, backend):
 
         tmp_output.seek(0)
         with laspy.open(tmp_output, closefd=False) as reader:
-            check_chunked_reading_is_gives_expected_points(
+            check_chunked_reading_gives_expected_points(
                 original_las, reader, iter_size
             )
-
-
-def check_chunked_reading_is_gives_expected_points(
-    groundtruth_las: laspy.LasData, reader: laspy.LasReader, iter_size: int
-):
-    """Checks that the points read by the reader are the same as groundtruth points."""
-    assert groundtruth_las.point_format == reader.header.point_format
-    for i, points in enumerate(reader.chunk_iterator(iter_size)):
-        expected_points = groundtruth_las.points[i * iter_size : (i + 1) * iter_size]
-        for dim_name in points.array.dtype.names:
-            assert np.allclose(
-                expected_points[dim_name], points[dim_name]
-            ), f"{dim_name} not equal"
-
 
 @pytest.mark.parametrize("backend", laspy.LazBackend.detect_available() + (None,))
 def test_chunked_dimension_modification(file_path, backend):
@@ -111,4 +107,4 @@ def test_chunked_dimension_modification(file_path, backend):
     ground_truth_las.x += 10.0
 
     reader = laspy.open(mem_dest, laz_backend=backend)
-    check_chunked_reading_is_gives_expected_points(ground_truth_las, reader, 42)
+    check_chunked_reading_gives_expected_points(ground_truth_las, reader, 42)