+++ tests/test_reading_1_2.py
@@ -226,3 +226,4 @@ def test_seek_las(las_file_path):
 def test_seek_laz(laz_file_path, laz_backend):
     with laspy.open(laz_file_path, laz_backend=laz_backend) as reader:
         check_seeking_works(reader)
+